✨ feat(Logging.ts): add Logging class to handle configuration of the event logging module
This commit adds a new file `Logging.ts` which contains the `Logging` class. This class is responsible for configuring the event logging module in the Discord bot. It provides a `logging` method that can be used as a slash command to set the channel where event logs will be sent.

The `logging` method takes a `channel` parameter, which represents the channel where event logs will be sent. It also takes an `interaction` parameter, which represents the command interaction.

Inside the `logging` method, there are several checks and validations. It checks if the bot has the `SendMessages` permission in the specified channel. If not, it sends a message to the user requesting the permission to be granted. It also checks if the specified channel is a text channel. If not, it sends an error message.

If all checks pass, the method sets the hosting channel for the guild in the Keyv database using the `setDb` function. It then sends a confirmation message to the user indicating that event logs will now be sent to the specified channel.