✨ feat(Emit.ts): add new command to set the role users will be given upon joining
This commit adds a new command called `autorole` to the `Emit.ts` file. The `autorole` command allows users with the `ManageRoles` permission to set the role that new members will be given upon joining the server.

The command takes a required `role` parameter, which should be a valid role in the server. The command checks if the bot has the `ManageRoles` permission and if the specified role is higher than the bot's highest role and the command executor's highest role. If any of these conditions are not met, an appropriate error message is sent as a reply.

If all conditions are met, the command saves the role ID in the database using the `setDb` function and sends a confirmation message to the user indicating that new members will now automatically be given the specified role.